Due to size mismatch between size_t and what is expected to create a SPECIAL instance, we are now using a `uint32' for `count', `capacity' and `index' in `struct obj_array'.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@95926 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/C/run-time/include/rt_traverse.h
@@ -93,9 +93,9 @@ struct mstack {
 
 struct obj_array {
 	EIF_REFERENCE *area;	/* Area where objects are stored */
-	size_t count;				/* Number of inserted items */
-	size_t capacity;			/* Capacity of `area' */
-	size_t index;				/* Cursor position */
+	uint32 count;				/* Number of inserted items */
+	uint32 capacity;			/* Capacity of `area' */
+	uint32 index;				/* Cursor position */
 };
 
 #ifdef __cplusplus
